+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Update install script
+
 ## [1.0.14] - 2024-04-27
 
 ### Fixed
@@ -79,6 +85,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_test/compare/1.0.14...HEAD
 [1.0.14]: https://github.com/inlavigo/gg_test/compare/1.0.13...1.0.14
 [1.0.13]: https://github.com/inlavigo/gg_test/compare/1.0.12...1.0.13
 [1.0.11]: https://github.com/inlavigo/gg_test/compare/1.0.10...1.0.11